# Speed up coordinate representations using erfa routines