--- a/astropy/units/quantity_helper.py
+++ b/astropy/units/quantity_helper.py
@@ -13,6 +13,7 @@ from fractions import Fraction
 import numpy as np
 from .core import (UnitsError, UnitConversionError, UnitTypeError,
                    dimensionless_unscaled, get_current_unit_registry)
+from .._erfa import ufunc as erfa_ufunc
 
 
 def _d(unit):
@@ -442,6 +443,45 @@ UFUNC_HELPERS[np.floor_divide] = helper_twoarg_floor_divide
 UFUNC_HELPERS[np.heaviside] = helper_heaviside
 UFUNC_HELPERS[np.float_power] = helper_power
 UFUNC_HELPERS[np.divmod] = helper_divmod
+
+
+# ERFA UFUNCS
+def helper_s2c(f, unit1, unit2):
+    from .si import radian
+    try:
+        return [get_converter(unit1, radian),
+                get_converter(unit2, radian)], dimensionless_unscaled
+    except UnitsError:
+        raise UnitTypeError("Can only apply '{0}' function to "
+                            "quantities with angle units"
+                            .format(f.__name__))
+
+
+def helper_s2p(f, unit1, unit2, unit3):
+    from .si import radian
+    try:
+        return [get_converter(unit1, radian),
+                get_converter(unit2, radian), None], unit3
+    except UnitsError:
+        raise UnitTypeError("Can only apply '{0}' function to "
+                            "quantities with angle units"
+                            .format(f.__name__))
+
+
+def helper_c2s(f, unit1):
+    from .si import radian
+    return [None], (radian, radian)
+
+
+def helper_p2s(f, unit1):
+    from .si import radian
+    return [None], (radian, radian, unit1)
+
+
+UFUNC_HELPERS[erfa_ufunc.s2c] = helper_s2c
+UFUNC_HELPERS[erfa_ufunc.s2p] = helper_s2p
+UFUNC_HELPERS[erfa_ufunc.c2s] = helper_c2s
+UFUNC_HELPERS[erfa_ufunc.p2s] = helper_p2s
 
 
 # UFUNCS FROM SCIPY.SPECIAL