--- a/astropy/units/quantity_helper.py
+++ b/astropy/units/quantity_helper.py
@@ -482,6 +482,7 @@ UFUNC_HELPERS[erfa_ufunc.s2c] = helper_s2c
 UFUNC_HELPERS[erfa_ufunc.s2p] = helper_s2p
 UFUNC_HELPERS[erfa_ufunc.c2s] = helper_c2s
 UFUNC_HELPERS[erfa_ufunc.p2s] = helper_p2s
+UFUNC_HELPERS[erfa_ufunc.pm] = helper_invariant
 
 
 # UFUNCS FROM SCIPY.SPECIAL