Follow-up of #7638, using `erfa` ufuncs for conversion of cartesian to (unit)spherical, and some of the cartesian methods.

Sample speed-up:
```
import numpy as np
from astropy.coordinates import CartesianRepresentation, SphericalRepresentation
cr = CartesianRepresentation(np.ones((100, 3)), unit=u.km, xyz_axis=-1)
%timeit cr.represent_as(SphericalRepresentation)
# 1000 loops, best of 3: 200 -> 174 µs
sph = cr.represent_as(SphericalRepresentation)
%timeit sph.represent_as(CartesianRepresentation)
# 1000 loops, best of 3: 233 -> 67 µs
%timeit cr.norm()
# 1000 loops, best of 3: 227 -> 9 µs
%timeit cr.dot(cr)
# 1000 loops, best of 3: 206 -> 28 µs
%timeit cr.cross(cr)
# 1000 loops, best of 3: 447 -> 66 µs
```