@adrn 2018-07-25
I played around with this and did some verification that the ERFA functions are indeed very fast (compared even to numpy). 

One thing I don't like about this PR is that, in some cases, it hides some fairly simple Python code into obscurely-named ERFA functions that makes it much less readable. That's mainly just an observation and something I'd like to discuss, rather than something that needs to be fixed in the PR, but I'm interested to hear your thoughts on solutions for readability. Maybe we should add better comments to the code that explain the ERFA functions (as you do in some places)?