@mhvk 2018-07-25
@adrn - I agree with the part that this makes the code less readable, and am even fairly sure I commented myself pointing that out on an earlier suggestion to wrap these functions in C! (not #5735, not #4664, even though that is being "fixed" here... just cannot seem to find it!).

My current sense is that using anything from `erfa` inside astropy is fair game, just as I wouldn't hesitate to use a possibly obscure numpy function. That said, erfa functions are arguably harder to look up, especially since we keep them in a private module, so adding comments is reasonable.

Also cc @eteq, given #4664