@mhvk 2018-07-25
@adrn - since I had to rebase anyway, I went ahead and added simple one-line comments for each use of an erfa ufunc (basically, just taking the first line of their docstrings).