@astrofrog 2018-08-04
This looks nice! Though @eteq and/or @adrn should review/approve