@astrofrog 2018-08-04
What we could consider doing in future is just create aliases for the ERFA functions with better names