@mhvk 2018-08-04
> What we could consider doing in future is just create aliases for the ERFA functions with better names

I thought about it, but in the end felt it would just be confusing (`Time` used to do that). I think adding comments was the right thing to do - this will be even more important to do inside `builtin_frames` - things like `atoq` are truly obscure!